pipeline_layer/requirements.txt
PyYAML
pandas
numpy
paho-mqtt
torch
ultralytics